update lsof: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@48112 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/lsof/tests/LTdnlc.c
@@ -186,7 +186,7 @@ main(argc, argv)
     }
     if ((em = ConvStatDev(&sb.st_dev, &cwddc)))
 	PrtMsgX(em, Pn, cleanup, 1);
-    (void) snprintf(ibuf, sizeof(ibuf) - 1, "%u", (unsigned int)sb.st_ino);
+    (void) snprintf(ibuf, sizeof(ibuf) - 1, "%" PRIu64, (uint64_t)sb.st_ino);
     ibuf[sizeof(ibuf) - 1] = '\0';
 /*
  * Loop ATTEMPT_CT times.
